refactor(admin): use clean architecture for product admin
