+++ .tool-versions
@@ -1,1 +1,1 @@
-golang 1.22.0
+golang 1.23.4
