--- internal/application/dto/product_output.go
@@ -1,10 +1,0 @@
-package dto
-
-type ProductOutput struct {
-	ID          int      `json:"id"`
-	Name        string   `json:"name"`
-	Price       float64  `json:"price"`
-	Description string   `json:"description"`
-	Category    string   `json:"category"`
-	Images      []string `json:"images"`
-}
