+++ internal/core/usecase/ports/product_repository.go
@@ -14,4 +14,9 @@ type ProductFilter struct {
 
 type ProductRepository interface {
 	GetAll(ctx context.Context, filter *ProductFilter) ([]domain.Product, int, error)
+	Create(ctx context.Context, product domain.Product) (domain.Product, error)
+	GetById(ctx context.Context, id int) (domain.Product, error)
+	Delete(ctx context.Context, id int) error
+	Update(ctx context.Context, product domain.Product) (domain.Product, error)
+	GetByIds(ctx context.Context, ids []int) ([]domain.Product, int, error)
 }
