--- internal/domain/entity/product.go
@@ -1,29 +1,0 @@
-package entity
-
-import "time"
-
-type ProductImage struct {
-	ID        int
-	ImageURL  string
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
-
-type ProductCategory struct {
-	ID        int
-	Name      string
-	Handle    string
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
-
-type Product struct {
-	ID          int
-	Name        string
-	Description string
-	Price       float64
-	Category    ProductCategory
-	Images      []ProductImage
-	CreatedAt   time.Time
-	UpdatedAt   time.Time
-}
