+++ internal/domain/repository.go
@@ -12,23 +12,9 @@ type UserRepository interface {
 	Create(ctx context.Context, user entity.User) (entity.User, error)
 }
 
-type ProductFilter struct {
-	Category string
-	Page     int
-	PageSize int
-}
-
 type OrderFilter struct {
 	Page     int
 	PageSize int
-}
-
-type ProductRepository interface {
-	Create(ctx context.Context, product entity.Product) (entity.Product, error)
-	GetById(ctx context.Context, id int) (entity.Product, error)
-	Delete(ctx context.Context, id int) error
-	Update(ctx context.Context, product entity.Product) (entity.Product, error)
-	GetByIds(ctx context.Context, ids []int) ([]entity.Product, int, error)
 }
 
 type OrderRepository interface {
